Move PublishOppgavehendelseCronjob to correct package
